+++ lib/karafka/web/pro/ui/routes/explorer.rb
@@ -47,56 +47,66 @@ module Karafka
                   end
                 end
 
-                controller = build(Controllers::Explorer::ExplorerController)
+                r.on 'topics' do
+                  controller = build(Controllers::Explorer::ExplorerController)
 
-                r.get String, :partition_id, 'recent' do |topic_id, partition_id|
-                  controller.recent(topic_id, partition_id)
-                end
+                  r.get String, :partition_id, 'recent' do |topic_id, partition_id|
+                    controller.recent(topic_id, partition_id)
+                  end
 
-                r.get(
-                  String, :partition_id, Integer, 'surrounding'
-                ) do |topic_id, partition_id, offset|
-                  controller.surrounding(topic_id, partition_id, offset)
-                end
+                  r.get(
+                    String, :partition_id, Integer, 'surrounding'
+                  ) do |topic_id, partition_id, offset|
+                    controller.surrounding(topic_id, partition_id, offset)
+                  end
 
-                r.get String, 'recent' do |topic_id|
-                  controller.recent(topic_id, nil)
-                end
+                  r.get String, 'recent' do |topic_id|
+                    controller.recent(topic_id, nil)
+                  end
 
-                # Jumps to offset matching the expected time
-                r.get String, :partition_id, 'closest', Time do |topic_id, partition_id, time|
-                  controller.closest(topic_id, partition_id, time)
-                end
+                  # Jumps to offset matching the expected time
+                  r.get String, :partition_id, 'closest', Time do |topic_id, partition_id, time|
+                    controller.closest(topic_id, partition_id, time)
+                  end
 
-                # Jumps to the offset matching the expected timestamp
-                r.get(
-                  String, :partition_id, 'closest', Integer
-                ) do |topic_id, partition_id, timestamp|
-                  # To simplify we just convert timestamp to time with ms precision
-                  time = Time.at(timestamp / 1_000.0)
-                  controller.closest(topic_id, partition_id, time)
-                end
+                  # Jumps to the offset matching the expected timestamp
+                  r.get(
+                    String, :partition_id, 'closest', Integer
+                  ) do |topic_id, partition_id, timestamp|
+                    # To simplify we just convert timestamp to time with ms precision
+                    time = Time.at(timestamp / 1_000.0)
+                    controller.closest(topic_id, partition_id, time)
+                  end
 
-                r.get String, :partition_id, Integer do |topic_id, partition_id, offset|
-                  # If when viewing given message we get an offset of different message, we should
-                  # redirect there. This allows us to support pagination with the current engine
-                  if params.current_offset != -1
-                    r.redirect explorer_path(topic_id, partition_id, params.current_offset)
-                  else
-                    controller.show(topic_id, partition_id, offset)
+                  r.get String, :partition_id, Integer do |topic_id, partition_id, offset|
+                    # If when viewing given message we get an offset of different message, we should
+                    # redirect there. This allows us to support pagination with the current engine
+                    if params.current_offset != -1
+                      r.redirect explorer_topics_path(
+                        topic_id,
+                        partition_id,
+                        params.current_offset
+                      )
+                    else
+                      controller.show(topic_id, partition_id, offset)
+                    end
+                  end
+
+                  r.get String, :partition_id do |topic_id, partition_id|
+                    controller.partition(topic_id, partition_id)
+                  end
+
+                  r.get String do |topic_id|
+                    controller.topic(topic_id)
+                  end
+
+                  r.get do
+                    controller.index
                   end
                 end
 
-                r.get String, :partition_id do |topic_id, partition_id|
-                  controller.partition(topic_id, partition_id)
-                end
-
-                r.get String do |topic_id|
-                  controller.topic(topic_id)
-                end
-
                 r.get do
-                  controller.index
+                  r.redirect root_path('explorer/topics')
                 end
               end
             end