+++ spec/spec_helper.rb
@@ -4,6 +4,9 @@ require 'factory_bot'
 require 'simplecov'
 require 'rack/test'
 require 'ostruct'
+require 'set'
+require 'nokogiri'
+require 'singleton'
 
 # Are we running regular specs or pro specs
 SPECS_TYPE = ENV.fetch('SPECS_TYPE', 'default')
@@ -84,6 +87,17 @@ RSpec.configure do |config|
   config.after(:suite) do
     PRODUCERS.regular.close
     PRODUCERS.transactional.close
+  end
+
+  config.after(:each, type: :controller) do |example|
+    # Do not proceed if there were any errors in the test
+    next if example.exception
+    # Analyze only valid html responses data
+    next unless response&.content_type&.include?('text/html')
+
+    validator = LinksValidator.instance
+    validator.context = self
+    validator.validate_all!(response)
   end
 end